requirements-minimal.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
alembic
fastapi-sqlalchemy
pydantic-settings
pydantic
passlib==1.7.4
python-jose
python-multipart
python-dotenv
jinja2
